+++ tests/EngineTest.php
@@ -27,6 +27,7 @@ use RoachPHP\Scheduling\ArrayRequestScheduler;
 use RoachPHP\Scheduling\Timing\FakeClock;
 use RoachPHP\Spider\ParseResult;
 use RoachPHP\Spider\Processor;
+use RoachPHP\Testing\Concerns\InteractsWithRequestsAndResponses;
 use RoachPHP\Testing\FakeLogger;
 use Symfony\Component\EventDispatcher\EventDispatcher;
