+++ tests/Http/RequestTest.php
@@ -19,7 +19,7 @@ use PHPUnit\Framework\TestCase;
 use RoachPHP\Http\Response;
 use RoachPHP\Spider\ParseResult;
 use RoachPHP\Support\DroppableInterface;
-use RoachPHP\Tests\InteractsWithRequestsAndResponses;
+use RoachPHP\Testing\Concerns\InteractsWithRequestsAndResponses;
 use RoachPHP\Tests\Support\DroppableTest;
 
 /**
